`Communication`: Add thumbnails for attached images below text field (#391)
* Restructure SendMessageViewModel

* Add image thumbnails

* Fix swiftlint warning

* Remove images after sending message

* Add line limit to image caption

* Decrease max height of text field
