+++ ArtemisKit/Sources/Messages/ViewModels/SendMessageViewModels/SendMessageViewModel.swift
@@ -45,6 +45,7 @@ final class SendMessageViewModel {
     // MARK: Text
 
     var text = ""
+    internal var uploadedImages = [String: UIImage]()
     internal var _selection: TextSelection? // swiftlint:disable:this identifier_name
     var selection: Binding<TextSelection?> {
         Binding {
@@ -176,102 +177,6 @@ extension SendMessageViewModel {
         }
     }
 
-    // MARK: Toolbar
-
-    func didTapBoldButton() {
-        appendToSelection(before: "**", after: "**", placeholder: "bold")
-    }
-
-    func didTapItalicButton() {
-        appendToSelection(before: "*", after: "*", placeholder: "italic")
-    }
-
-    func didTapUnderlineButton() {
-        appendToSelection(before: "<ins>", after: "</ins>", placeholder: "underlined")
-    }
-
-    func didTapStrikethroughButton() {
-        appendToSelection(before: "~~", after: "~~", placeholder: "strikethough")
-    }
-
-    func didTapBlockquoteButton() {
-        appendToSelection(before: "> ", after: "", placeholder: "Reference")
-    }
-
-    func didTapCodeButton() {
-        appendToSelection(before: "`", after: "`", placeholder: "Code")
-    }
-
-    func didTapCodeBlockButton() {
-        appendToSelection(before: "```java\n", after: "\n```", placeholder: "Source Code")
-    }
-
-    func didTapLinkButton() {
-        appendToSelection(before: "[", after: "](https://)", placeholder: "Display Text")
-    }
-
-    func didTapAtButton() {
-        if conditionalPresentation == .memberPicker {
-            isMemberPickerSuppressed = true
-        } else {
-            isMemberPickerSuppressed = false
-            appendToSelection(before: "@", after: " ", placeholder: " ")
-        }
-    }
-
-    func didTapNumberButton() {
-        if conditionalPresentation == .channelPicker {
-            isChannelPickerSuppressed = true
-        } else {
-            isChannelPickerSuppressed = false
-            appendToSelection(before: "#", after: " ", placeholder: " ")
-        }
-    }
-
-    func insertImageMention(path: String) {
-        appendToSelection(before: "![", after: "](\(path))", placeholder: "image")
-    }
-
-    func insertFileMention(path: String, fileName: String) {
-        appendToSelection(before: "[", after: "](\(path))", placeholder: fileName)
-    }
-
-    /// Prepends/Appends the given snippets to text the user has selected.
-    private func appendToSelection(before: String, after: String, placeholder: String) {
-        let placeholderText = "\(before)\(placeholder)\(after)"
-        var shouldSelectPlaceholder = false
-
-        if let selection = _selection {
-            switch selection.indices {
-            case .selection(let range):
-                let newText: String
-                if text[range].isEmpty {
-                    newText = placeholderText
-                    shouldSelectPlaceholder = true
-                } else {
-                    newText = "\(before)\(text[range])\(after)"
-                }
-                text.replaceSubrange(range, with: newText)
-                if !shouldSelectPlaceholder {
-                    moveCursor(after: newText)
-                }
-            default:
-                break
-            }
-        } else {
-            text.append(placeholderText)
-            shouldSelectPlaceholder = true
-        }
-
-        if shouldSelectPlaceholder {
-            for range in text.ranges(of: placeholderText) {
-                if let placeholderRange = text[range].range(of: placeholder) {
-                    _selection = TextSelection(range: range.clamped(to: placeholderRange))
-                }
-            }
-        }
-    }
-
     // MARK: Send Message
 
     func didTapSendButton() {
@@ -306,6 +211,7 @@ extension SendMessageViewModel {
             case .success:
                 _selection = nil
                 text = ""
+                uploadedImages = [:]
             default:
                 return
             }
@@ -321,6 +227,7 @@ extension SendMessageViewModel {
         case .success:
             await completion()
             isLoading = false
+            uploadedImages = [:]
         case .failure(let error):
             isLoading = false
             if let apiClientError = error as? APIClientError {
@@ -339,6 +246,7 @@ extension SendMessageViewModel {
         case .notStarted, .loading:
             return false
         case .success:
+            uploadedImages = [:]
             return true
         case .failure(let error):
             delegate.presentError(UserFacingError(title: error.localizedDescription))
@@ -351,50 +259,12 @@ extension SendMessageViewModel {
 
         switch result {
         case .notStarted, .loading: return false
-        case .success: return true
+        case .success:
+            uploadedImages = [:]
+            return true
         case .failure(let error):
             delegate.presentError(UserFacingError(title: error.localizedDescription))
             return false
-        }
-    }
-
-    // MARK: Search and Replace
-    func searchChannel() -> Substring? {
-        return text.matches(of: #/#(?<candidate>[\w-]*)/#).last?.candidate
-    }
-
-    func replace(channel: ChannelIdAndNameDTO) {
-        guard let candidate = searchChannel() else {
-            return
-        }
-        let channelMention = "[channel]\(channel.name)(\(channel.id))[/channel]"
-        text = text.replacingOccurrences(of: "#" + candidate, with: channelMention)
-        moveCursor(after: channelMention)
-    }
-
-    func searchMember() -> Substring? {
-        let matches = text.matches(of: #/@(?<candidate>[\w]*\s?)/#)
-        let candidate = matches.last?.candidate
-        if candidate == " " || candidate?.contains(".") == true {
-            // Either space after @ or a period indicating an email address -> Hide the member picker
-            return nil
-        }
-        return candidate?.filter { $0 != Character(" ") }
-    }
-
-    func replace(member: UserNameAndLoginDTO) {
-        guard let candidate = searchMember(),
-              let name = member.name, let login = member.login else {
-            return
-        }
-        let userMention = "[user]\(name)(\(login))[/user]"
-        text = text.replacingOccurrences(of: "@" + candidate, with: userMention)
-        moveCursor(after: userMention)
-    }
-
-    private func moveCursor(after string: String) {
-        if let range = text.range(of: string) {
-            _selection = TextSelection(insertionPoint: range.upperBound)
         }
     }
 }
